requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
